requirements.txt
# ── Core HTTP ─────────────────────────────────────────────────────────────────
httpx==0.27.0           # async HTTP client (REST + GraphQL)

# ── Database ──────────────────────────────────────────────────────────────────
asyncpg==0.29.0         # async PostgreSQL driver

# ── Cache ─────────────────────────────────────────────────────────────────────
redis[asyncio]==5.0.4   # Redis client

# ── Data processing ───────────────────────────────────────────────────────────
duckdb==1.0.0           # in-process OLAP (analytics layer, Spiral 2)
pandas==2.2.2           # DataFrames for ETL transforms
pyarrow==16.1.0         # Parquet read/write

# ── Analytics / anomaly detection (Spiral 2) ──────────────────────────────────
scikit-learn==1.5.0     # Isolation Forest, LOF
numpy==1.26.4           # numerical base

# ── Vector embeddings (Spiral 3) ─────────────────────────────────────────────
sentence-transformers==3.0.0   # multilingual embeddings (RU+KZ)
pymilvus==2.4.3                # Milvus vector store client

# ── API layer (Spiral 2) ──────────────────────────────────────────────────────
fastapi==0.111.0        # API framework
uvicorn[standard]==0.30.1
streamlit==1.38.0       # demo UI

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1    # load .env files
pydantic==2.7.4         # data validation / settings
tenacity==8.3.0         # retry decorators
python-dateutil==2.9.0  # date parsing
tqdm==4.66.4            # progress bars for ETL

# ── Dev / test ────────────────────────────────────────────────────────────────
pytest==8.2.2
pytest-asyncio==0.23.7
ruff==0.4.8             # linter + formatter
